move graph.py and jit.py into features (#3376)
* move graph.py into features

* move jit into features

* fix quickstart
