+++ docs/abstractions2.py
@@ -91,7 +91,7 @@ sched = out.schedule()
 for si in sched: print(si.ast.op)  # NOTE: the first two convert it to CLANG
 
 # DEBUGGING: print the compute ast as a tree
-from tinygrad.graph import print_tree
+from tinygrad.features.graph import print_tree
 print_tree(sched[-1].ast)
 # NOTE: sched[-1].ast is the same as st_0 above
