+++ examples/benchmark_train_efficientnet.py
@@ -7,7 +7,7 @@ from tinygrad.nn.state import get_parameters
 from tinygrad.nn import optim
 from tinygrad import Tensor, GlobalCounters
 from tinygrad.helpers import getenv
-from tinygrad.jit import CacheCollector
+from tinygrad.features.jit import CacheCollector
 
 def tensors_allocated():
   return sum(isinstance(x, Tensor) for x in gc.get_objects())
