+++ examples/stable_diffusion.py
@@ -13,7 +13,7 @@ from tinygrad import Device, GlobalCounters, dtypes, Tensor
 from tinygrad.helpers import Timing, Context, getenv, fetch, colored
 from tinygrad.nn import Conv2d, Linear, GroupNorm, LayerNorm, Embedding
 from tinygrad.nn.state import torch_load, load_state_dict, get_state_dict
-from tinygrad.jit import TinyJit
+from tinygrad.features.jit import TinyJit
 
 class AttnBlock:
   def __init__(self, in_channels):
