+++ test/test_symbolic_jit.py
@@ -1,7 +1,7 @@
 import unittest
 
 from test.helpers import assert_jit_cache_len
-from tinygrad.jit import TinyJit
+from tinygrad.features.jit import TinyJit
 from tinygrad.helpers import getenv
 from tinygrad.shape.symbolic import Variable
 from tinygrad.tensor import Tensor
