+++ tinygrad/__init__.py
@@ -1,5 +1,5 @@
 from tinygrad.tensor import Tensor            # noqa: F401
-from tinygrad.jit import TinyJit              # noqa: F401
+from tinygrad.features.jit import TinyJit     # noqa: F401
 from tinygrad.shape.symbolic import Variable  # noqa: F401
 from tinygrad.dtype import dtypes             # noqa: F401
 from tinygrad.ops import GlobalCounters       # noqa: F401
